+++ NEWS.md
@@ -1,3 +1,4 @@
+* Stop loading the entire namespace of a suggested package when it is available #296
 * Unexport objects: `.import`, `.export`; remove documentation for `arg_reconcile` #321
 * Declutter
    - remove the obsolete data.table option #323